GroundStation_AeroDock/requirements.txt
PyQt5>=5.15.0
PyOpenGL>=3.1.0
pyserial>=3.5
numpy>=1.19.0
hidapi>=0.14.0
